Done Gallery-Gallerybackend

+++ app/Http/Controllers/GalleryController.php
@@ -37,8 +37,8 @@ class GalleryController extends Controller
         $gallery = Gallery::where('product_id', $product_id)->get();
         $gallery_count = $gallery->count();
         $output = '
-        <form>
-                ' . csrf_field() . '
+    <form>
+        ' . csrf_field() . '
         <table class="table table-hover">
             <thead>
                 <tr>
@@ -49,54 +49,53 @@ class GalleryController extends Controller
                 </tr>
             </thead>
             <tbody>
-    ';
+';
 
         if ($gallery_count > 0) {
             $i = 0;
             foreach ($gallery as $key => $gal) {
                 $i++;
                 $output .= '
-            
-                <tr>
-                    <td>' . $i . '</td>
-                    <td contenteditable class="edit_gal_name" data-gal_id="' . $gal->gallery_id . '">' . $gal->gallery_name . '</td>
-                    <td><img src = "' . url('uploads/gallery/' . $gal->gallery_image) . '" class="img-thumbnail" width="120px" height="120px"></td>
-                    <td><button type="button" data-gal_id="' . $gal->gallery_id . '" class="btn btn-xs btn-danger delete-gallery">Xóa</button></td>
-                    
-                </tr>
-            ';
+            <tr>
+                <td>' . $i . '</td>
+                <td contenteditable class="edit_gal_name" data-gal_id="' . $gal->gallery_id . '">' . $gal->gallery_name . '</td>
+                <td>
+                    <img src="' . url('uploads/gallery/' . $gal->gallery_image) . '" class="img-thumbnail" width="120px" height="120px">
+                    <input type="file" class="file_image" style="width:40%" data-gal_id="' . $gal->gallery_id . '" id="file-' . $gal->gallery_id . '" name="file" accept="image/*"/>
+                </td>
+                <td><button type="button" data-gal_id="' . $gal->gallery_id . '" class="btn btn-xs btn-danger delete-gallery">Xóa</button></td>
+            </tr>
+        ';
             }
         } else {
             $output .= '
-            <tr>
-                <td colspan="4">Sản phẩm này chưa có thư viện ảnh</td>
-            </tr>
+        <tr>
+            <td colspan="4">Sản phẩm này chưa có thư viện ảnh</td>
+        </tr>
         ';
         }
+
         $output .= '
             </tbody>
             </table>
             </form>
         ';
 
+
         echo $output;
     }
 
     public function insert_gallery(Request $request, $pro_id)
     {
-        $request->validate([
-            'file' => 'required|array',
-            'file.*' => 'image|mimes:jpeg,png,jpg,gif,svg|max:2048'
-        ]);
-
-        $get_image = $request->file('file');
-        if ($get_image) {
-            foreach ($get_image as $image) {
+        $images = $request->file('file');
+        if ($images) {
+            foreach ($images as $image) {
                 $get_name_image = $image->getClientOriginalName();
-                $name_image = current(explode('.', $get_name_image));
-                $new_image = $get_name_image . rand(0, 99) . '.' . $image->getClientOriginalExtension();
-
+                $name_image = pathinfo($get_name_image, PATHINFO_FILENAME);
+                $new_image = $name_image . rand(0, 99) . '.' . $image->getClientOriginalExtension();
+    
                 $image->move('uploads/gallery', $new_image);
+    
                 $gallery = new Gallery();
                 $gallery->gallery_name = $new_image;
                 $gallery->gallery_image = $new_image;
@@ -104,9 +103,11 @@ class GalleryController extends Controller
                 $gallery->save();
             }
         }
+    
         Session::put('message', 'Thêm thư viện ảnh thành công');
         return redirect()->back();
     }
+    
 
     public function update_gallery_name(Request $request)
     {
@@ -121,8 +122,25 @@ class GalleryController extends Controller
     {
         $gal_id = $request->gal_id;
         $gallery = Gallery::find($gal_id);
-        unlink('uploads/gallery/'.$gallery->gallery_image);
+        unlink('uploads/gallery/' . $gallery->gallery_image);
         $gallery->delete();
+    }
 
+    public function update_gallery(Request $request)
+    {
+        $get_image = $request->file('file');
+        $gal_id = $request->gal_id;
+        if ($get_image) {
+            $get_name_image = $get_image->getClientOriginalName();
+            $name_image = current(explode('.', $get_name_image));
+            $new_image = $get_name_image . rand(0, 99) . '.' . $get_image->getClientOriginalExtension();
+
+            $get_image->move('uploads/gallery', $new_image);
+            $gallery =  Gallery::find($gal_id);
+            unlink('uploads/gallery/' . $gallery->gallery_image);
+
+            $gallery->gallery_image = $new_image;
+            $gallery->save();
+        }
     }
 }
